Implement support for ifHighSpeed OID
Until this commit we were using ifSpeed OID to find out interface speed
(bitrate). This OID does not support large speeds over 40G, so we needed
to switch over to ifHighSpeed. If ifHighSpeed is present, it is preferred
to ifSpeed. This also necessitated moving the 'status.rate' backend field
from smallint to bigint.

+++ schema/status.sql
@@ -15,7 +15,7 @@ CREATE TABLE status (
   vlans       bit(4096),
   descr       character varying(64),
   duplex      smallint,
-  rate        smallint,
+  rate        bigint,
   flags       smallint,
   adminstatus boolean,
   errdis      boolean,
